Inital working version
