+++ stixpy/products/__init__.py
@@ -1,0 +1,1 @@
+from stixpy.products.product_factory import Product
