+++ stixpy/products/tests/test_product_factory.py
@@ -1,0 +1,13 @@
+from stixpy.products.product_factory import Product
+
+from stixpy.science import *
+
+# def test_product_factory_file():
+#     p = Product('/Users/shane/sunpy/data/solo_L1_stix-sci-xray-cpd_20221115T023650-20221115T033526_V01_2211153759-54903.fits')
+#     assert False
+#
+# def test_product_factory_url():
+#     assert False
+#
+# def test_product_factory_meta_qtables():
+#     assert False
